# Conflicts:
#	src/command/restore/restore.c

+++ src/command/backup/backup.c
@@ -1195,9 +1195,9 @@ backupJobResult(
                 const String *fileLog = host == NULL ? fileName : strNewFmt("%s:%s", strZ(host), strZ(fileName));
 
                 // Format log strings
-                const String *const logProgress =
-                    strNewFmt(
-                        "%s, %" PRIu64 "%%", strZ(strSizeFormat(copySize)), sizeTotal == 0 ? 100 : *sizeProgress * 100 / sizeTotal);
+                const String *const logProgress = strNewFmt(
+                    "%s, %.2lf%%", strZ(strSizeFormat(copySize)),
+                    sizeTotal == 0 ? 100.00 : (double)*sizeProgress * 100.0 / (double)sizeTotal);
                 const String *const logChecksum = copySize != 0 ? strNewFmt(" checksum %s", strZ(copyChecksum)) : EMPTY_STR;
 
                 // If the file is in a prior backup and nothing changed, just log it
